+++ src/lib/output/themes/default/DefaultTheme.tsx
@@ -17,6 +17,7 @@ import type { MarkedPlugin } from "../../plugins";
 import { DefaultThemeRenderContext } from "./DefaultThemeRenderContext";
 import { JSX } from "../../../utils";
 import { classNames, getDisplayName, getHierarchyRoots, toStyleClass } from "../lib";
+import { icons } from "./partials/icon";
 
 /**
  * Defines a mapping of a {@link Models.Kind} to a template file.
@@ -55,6 +56,21 @@ export interface NavigationElement {
 export class DefaultTheme extends Theme {
     /** @internal */
     markedPlugin: MarkedPlugin;
+
+    /**
+     * The icons which will actually be rendered. The source of truth lives on the theme, and
+     * the {@link DefaultThemeRenderContext.icons} member will produce references to these.
+     *
+     * These icons will be written twice. Once to an `icons.svg` file in the assets directory
+     * which will be referenced by icons on the context, and once to an `icons.js` file so that
+     * references to the icons can be dynamically embedded within the page for use by the search
+     * dropdown and when loading the page on `file://` urls.
+     *
+     * Custom themes may overwrite this entire object or individual properties on it to customize
+     * the icons used within the page, however TypeDoc currently assumes that all icons are svg
+     * elements, so custom themes must also use svg elements.
+     */
+    icons = { ...icons };
 
     getRenderContext(pageEvent: PageEvent<Reflection>) {
         return new DefaultThemeRenderContext(this, pageEvent, this.application.options);